+++ src/main/java/JesusDeciples/RankingQuiz/api/webSocket/QuizDataCenter.java
@@ -7,8 +7,8 @@ import JesusDeciples.RankingQuiz.api.dto.response.QuizResultDto;
 import JesusDeciples.RankingQuiz.api.entity.quiz.Quiz;
 import JesusDeciples.RankingQuiz.api.facade.QuizQuizContentFacade;
 import JesusDeciples.RankingQuiz.api.facade.QuizScoreFacade;
-import JesusDeciples.RankingQuiz.api.service.quizDataCenter.DataCenterState;
-import JesusDeciples.RankingQuiz.api.service.quizDataCenter.WAITING;
+import JesusDeciples.RankingQuiz.api.service.quizDataCenter.state.DataCenterState;
+import JesusDeciples.RankingQuiz.api.service.quizDataCenter.state.WAITING;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
